+++ docs/NEWS.md
@@ -1,6 +1,7 @@
 ### UNRELEASED
 
  * *fix:* add _configure_ script fallback for **--remote** check when cross-compiling (#120)
+ * *fix:* allow **extra-display** to be changed by **--remote** (#123)
 
 ### 1.9.7 - 2 December 2024